Visualisation improved
Wrote functions for graphing and tidied up code. Added a line of best fit, but that only worked if I reduced the time variable significantly. It currently starts at zero which is pretty unphysical, but maybe I can have it to reset to zero every day? That would mean creating a graph per calendar day?