Generation of readings
Wrote a function to generate a value for sensor_outputs every second, which is dictated by the first value given in the example and shifted by a random amount dictated by the Normal distribution. The number of significant figures for each quantity is kept.
The storing of readings is turned into a function for modulisation.
Re-naming of variables to ease reading.